# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Orange Money",
        "description": "API pour le système de mobile money Orange Money",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes de l'utilisateur connecté",
                "operationId": "getComptes",
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "operationId": "POST::/comptes::App\\Http\\Controllers\\CompteController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_compte",
                                    "titulaire",
                                    "statut"
                                ],
                                "properties": {
                                    "numero_compte": {
                                        "type": "string",
                                        "example": "CMPT-001"
                                    },
                                    "titulaire": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "code_marchand": {
                                        "type": "string",
                                        "example": "MRC001"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif",
                                            "suspendu"
                                        ],
                                        "example": "actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comptes/{compte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher un compte spécifique",
                "operationId": "GET::/comptes/{compte}::App\\Http\\Controllers\\CompteController::show",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte",
                "operationId": "PUT::/comptes/{compte}::App\\Http\\Controllers\\CompteController::update",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulaire": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "code_marchand": {
                                        "type": "string",
                                        "example": "MRC002"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif",
                                            "suspendu"
                                        ],
                                        "example": "actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte (soft delete)",
                "operationId": "DELETE::/comptes/{compte}::App\\Http\\Controllers\\CompteController::destroy",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "otp_code": {
                                        "description": "Code OTP requis pour supprimer un compte actif",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/comptes/me": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher les comptes de l'utilisateur connecté",
                "operationId": "GET::/comptes/me::App\\Http\\Controllers\\CompteController::me",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Inclure tous les comptes (par défaut seulement actifs)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comptes récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre compte actif récupéré avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "numero_compte": {
                                                        "type": "string",
                                                        "example": "CMPT-001"
                                                    },
                                                    "titulaire": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "solde": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15000.5
                                                    },
                                                    "statut": {
                                                        "type": "string",
                                                        "example": "actif"
                                                    },
                                                    "code_marchand": {
                                                        "type": "string",
                                                        "example": "MRC001",
                                                        "nullable": true
                                                    },
                                                    "qr_code": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "uuid-compte"
                    },
                    "numero_compte": {
                        "type": "string",
                        "example": "CMPT-001"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "code_marchand": {
                        "type": "string",
                        "example": "MRC001",
                        "nullable": true
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "inactif",
                            "suspendu"
                        ],
                        "example": "actif"
                    },
                    "utilisateur_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "uuid-user"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 15000.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}